add github org plugin
