+++ dynamic-plugins/wrappers/backstage-plugin-catalog-backend-module-github-org-dynamic/.eslintignore
@@ -1,0 +1,1 @@
+dist-dynamic
