+++ dynamic-plugins/wrappers/backstage-plugin-catalog-backend-module-github-org-dynamic/.eslintrc.js
@@ -1,0 +1,1 @@
+module.exports = require('@backstage/cli/config/eslint-factory')(__dirname);
