+++ dynamic-plugins/wrappers/backstage-plugin-catalog-backend-module-github-org-dynamic/src/dynamic/alpha.ts
@@ -1,0 +1,8 @@
+import { BackendDynamicPluginInstaller } from '@backstage/backend-plugin-manager';
+import catalogModuleGithubOrgEntityProvider from '@backstage/plugin-catalog-backend-module-github-org';
+
+export const dynamicPluginInstaller: BackendDynamicPluginInstaller = {
+  kind: 'new',
+
+  install: () => [catalogModuleGithubOrgEntityProvider()],
+};
