+++ dynamic-plugins/wrappers/backstage-plugin-catalog-backend-module-github-org-dynamic/src/index.ts
@@ -1,0 +1,1 @@
+export * from './dynamic/index';
